scraper/requirements.txt
playwright==1.41.0
beautifulsoup4==4.12.3
requests==2.31.0
supabase==2.3.4
python-dotenv==1.0.0
twilio==8.12.0
schedule==1.2.1
